requirements.txt
flask
nltk
requests
numpy
tensorflow
transformers
google-cloud-bigquery